% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiveR-package.R
\docType{package}
\name{DiveR-package}
\alias{DiveR}
\alias{DiveR-package}
\title{DiveR: Visualization Of Viral Protein Sequence Diversity Dynamics}
\description{
DiveR allow users to visualize the diversity motifs (index and its variants – major, minor and unique) for elucidation of the underlying inherent dynamics.
}
\keyword{internal}
