% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_extraction.R
\name{extract_from_GISAID}
\alias{extract_from_GISAID}
\title{Extract metadata via fasta file from GISAID}
\usage{
extract_from_GISAID(file_path)
}
\arguments{
\item{file_path}{path of fasta file}
}
\description{
This function get the metadata from each header of GISAID fasta file
}
