% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_extraction.R
\name{extract_from_NCBI}
\alias{extract_from_NCBI}
\title{Extract metadata via fasta file from ncbi}
\usage{
extract_from_NCBI(file_path)
}
\arguments{
\item{file_path}{path of fasta file}
}
\description{
This function get the metadata from each head of fasta file
}
