% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{JSON_sample}
\alias{JSON_sample}
\title{DiMA (v5.0.9) JSON Output File}
\format{
A Diversity Motif Analyzer (DiMA) tool JSON file
}
\usage{
JSON_sample
}
\description{
A sample DiMA JSON Output File which acts as the input for JSON2CSV()
}
\keyword{datasets}
