% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{JSONsample}
\alias{JSONsample}
\title{DiMA (v4.1.1) JSON Output File}
\format{
A Diversity Motif Analyzer (DiMA) tool JSON file
}
\usage{
JSONsample
}
\description{
A sample DiMA JSON Output File which acts as the input for JSON2CSV()
}
\keyword{datasets}
