% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_world_map.R
\name{match_region_to_target}
\alias{match_region_to_target}
\title{Map data region to ggplot2 region}
\usage{
match_region_to_target(region_column, region_list, max_dist = 3)
}
\arguments{
\item{region_column}{a character vector of data region}

\item{region_list}{a character vector of ggplot2 region}

\item{max_dist}{maximum distance for string matching}
}
\description{
This function maps the provided data region to ggplot2 region.
}
