% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata Input Sample}
\format{
A data frame with 1000 rows and 3 variables:
\describe{
  \item{ID}{unique identifier of the sequence}
  \item{region}{geographical region of the sequence collection}
  \item{date}{collection date of the sequence}
}
}
\usage{
metadata
}
\description{
A dummy dataset that acts as an input for plot_world_map() and plot_time()
}
\keyword{datasets}
