% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time.R
\name{plot_time}
\alias{plot_time}
\title{Time Distribution of Sequences Plot}
\usage{
plot_time(meta)
}
\arguments{
\item{meta}{a dataframe with 3 columns, 'ID', 'country', and 'date'}
}
\value{
A plot
}
\description{
This function plots the time distribution of provided sequences in the form of bar plot with 'Month' as x-axis and 
'sequence recorded' as y-axis.
}
\examples{
filepath <- system.file('extdata','NCBI_Protein.faa', package = 'vDiveR')
meta_ncbi <- metadata_extraction(filepath, 'ncbi')
plot_time(meta_ncbi)

filepath <- system.file('extdata','GISAID_EpiCoV.faa', package = 'vDiveR')
meta_gisaid <- metadata_extraction(filepath, 'GISAID')
plot_time(meta_gisaid)

plot_time(metadata)
}
