% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_worldmap.R
\name{plot_worldmap}
\alias{plot_worldmap}
\title{plot the geographical location via metadata}
\usage{
plot_worldmap(meta)
}
\arguments{
\item{meta}{sequence metadata from 'metadataExtraction' function or user provided, the format of metadata can refer to built-in data, "metadata".}
}
\value{
A plot
}
\description{
This function plot the cumulative sequence number in geographical via metadata
}
\examples{
filepath <- system.file('inst/extdata','prot_NCBIProtein.fasta', package = 'vDiveR')
meta <- metadataExtraction(filepath, 'ncbi')
plot_worldmap(meta)
plot_worldmap(metadata)
}
