% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiveR-package.R
\docType{package}
\name{vDiveR-package}
\alias{vDiveR}
\alias{vDiveR-package}
\title{vDiveR: Visualization of Viral Protein Sequence Diversity Dynamics}
\description{
To ease the visualization of outputs from Diversity Motif Analyser ('DiMA'; \url{https://github.com/BVU-BILSAB/DiMA}). 'vDiveR' allows visualization of the diversity motifs (index and its variants – major, minor and unique) for elucidation of the underlying inherent dynamics. Please refer \url{https://vdiver-manual.readthedocs.io/en/latest/} for more information.
}
\author{
\strong{Maintainer}: Pendy Tok \email{pendytok0518@gmail.com}

Authors:
\itemize{
  \item Li Chuin Chong
  \item Evgenia Chikina
  \item Yin Cheng Chen
  \item Mohammad Asif Khan
}

}
\keyword{internal}
