% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json2csv.R
\name{json2csv}
\alias{json2csv}
\title{JSON2CSV}
\usage{
json2csv(infile, hostName = "unknown host", proteinName = "unknown protein")
}
\arguments{
\item{infile}{DiMA JSON output file}

\item{hostName}{name of the host species}

\item{proteinName}{name of the protein}
}
\value{
A dataframe which acts as input for the other functions in DiveR package
}
\description{
This function converts DiMA (v4.1.1) JSON output file to a dataframe with 17
predefined columns which further acts as the input for other functions provided in this DiveR package.
}
\examples{
inputdf<-json2csv(JSONsample)
}
