% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dynamics_protein.R
\name{plot_dynamics_protein}
\alias{plot_dynamics_protein}
\title{Plot dynamics of diversity motifs (Protein)}
\usage{
plot_dynamics_protein(
  df,
  host = 1,
  proteinOrder = "",
  base_size = 8,
  alpha = 1/3,
  dot_size = 3
)
}
\arguments{
\item{df}{DiMA JSON converted csv file data}

\item{host}{number of host (1/2)}

\item{proteinOrder}{order of proteins displayed in plot}

\item{base_size}{base font size in plot}

\item{alpha}{any number from 0 (transparent) to 1 (opaque)}

\item{dot_size}{dot size in scatter plot}
}
\value{
A plot
}
\description{
Plot dynamics of diversity motifs (Protein)
}
\examples{
plot_dynamics_protein(proteins_1host)
plot_dynamics_protein(protein_2hosts, base_size = 8, dot_size = 3, alpha=0.1, host = 2)
}
