% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dynamics_protein.R
\name{plot_dynamics_protein}
\alias{plot_dynamics_protein}
\title{Dynamics of Diversity Motifs (Protein) Plot}
\usage{
plot_dynamics_protein(
  df,
  host = 1,
  proteinOrder = "",
  base_size = 8,
  alpha = 1/3,
  dot_size = 3
)
}
\arguments{
\item{df}{DiMA JSON converted csv file data}

\item{host}{number of host (1/2)}

\item{proteinOrder}{order of proteins displayed in plot}

\item{base_size}{base font size in plot}

\item{alpha}{any number from 0 (transparent) to 1 (opaque)}

\item{dot_size}{dot size in scatter plot}
}
\value{
A plot
}
\description{
This function compactly display the dynamics of diversity motifs (index and its variants: major, minor and unique)
in the form of dot plot(s) as well as violin plots for all the provided individual protein(s).
}
\examples{
plot_dynamics_protein(proteins_1host)
}
