% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dynamics_proteome.R
\name{plot_dynamics_proteome}
\alias{plot_dynamics_proteome}
\title{Dynamics of diversity motifs (Proteome) Plot}
\usage{
plot_dynamics_proteome(df, host = 1, dot_size = 2, word_size = 15, alpha = 1/3)
}
\arguments{
\item{df}{DiMA JSON converted csv file data}

\item{host}{number of host (1/2)}

\item{dot_size}{size of dot in plot}

\item{word_size}{word size in plot}

\item{alpha}{any number from 0 (transparent) to 1 (opaque)}
}
\value{
A plot
}
\description{
Plot dynamics of diversity motifs (Proteome)
}
\examples{
plot_dynamics_proteome(proteins_1host)
plot_dynamics_proteome(protein_2hosts, word_size = 8, dot_size = 3, alpha=0.1, host = 2)
}
