% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time.R
\name{plot_time}
\alias{plot_time}
\title{plot the time of collection via metadata}
\usage{
plot_time(meta)
}
\arguments{
\item{meta}{sequence metadata from 'metadata_extraction' function or user provided, the format of metadata can refer to built-in data, "metadata".}
}
\value{
A plot
}
\description{
This function plot the time of collection via metadata
}
\examples{
filepath <- system.file('inst/extdata','prot_NCBIProtein.fasta', package = 'vDiveR')
meta <- metadata_extraction(filepath, 'ncbi')
plot_time(meta)
plot_time(metadata)
}
