% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time.R
\name{plot_time}
\alias{plot_time}
\title{Time Distribution of Sequences Plot}
\usage{
plot_time(meta)
}
\arguments{
\item{meta}{a dataframe with 3 columns, 'ID', 'country', and 'date'}
}
\value{
A plot
}
\description{
This function plots the time distribution of provided sequences in the form of bar plot with 'Month' as x-axis and 
'sequence recorded' as y-axis. The input dataframe of this function is obtainable from metadata_extraction(), with NCBI 
Protein / GISAID EpiCoV FASTA file as input.
}
\examples{
plot_time(metadata)
}
