% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_worldmap.R
\name{plot_worldmap}
\alias{plot_worldmap}
\title{Geographical Distribution of Sequences Plot}
\usage{
plot_worldmap(meta)
}
\arguments{
\item{meta}{a dataframe with 3 columns, 'ID', 'country', and 'date'}
}
\value{
A plot
}
\description{
This function plots a worldmap and color the affected geographical region(s)
from light (lower) to dark (higher), depends on the cumulative number of sequences. 
The input dataframe of this function is obtainable from metadata_extraction(), with NCBI 
Protein / GISAID EpiCoV FASTA file as input.
}
\examples{
plot_worldmap(metadata)
}
