% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_worldmap.R
\name{plot_worldmap}
\alias{plot_worldmap}
\title{Geographical Distribution of Sequences Plot}
\usage{
plot_worldmap(meta)
}
\arguments{
\item{meta}{a dataframe with 3 columns, 'ID', 'country', and 'date'}
}
\value{
A plot
}
\description{
This function plots a worldmap and color the affected geographical region(s)
from light (lower) to dark (higher), depends on the cumulative number of sequences.
}
\examples{
filepath <- system.file('extdata','NCBI_Protein.faa', package = 'vDiveR')
meta_ncbi <- metadata_extraction(filepath, 'ncbi')
plot_worldmap(meta_ncbi)

filepath <- system.file('extdata','GISAID_EpiCoV.faa', package = 'vDiveR')
meta_gisaid <- metadata_extraction(filepath, 'GISAID')
plot_worldmap(meta_gisaid)

plot_worldmap(metadata)
}
