% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiveR-package.R
\docType{package}
\name{vDiveR-package}
\alias{vDiveR}
\alias{vDiveR-package}
\title{vDiveR: Visualization of Viral Protein Sequence Diversity Dynamics}
\description{
Allow users to visualize the diversity motifs (index and its variants – major, minor and unique) for elucidation of the underlying inherent dynamics.
}
\keyword{internal}
